requirements.txt
streamlit
beautifulsoup4
httpx
google-generativeai
pydantic-settings
python-dotenv
lxml
